Create _icons.scss
This is generally how we set stuff up with icons - we're usually working via icomoon, but the concepts here should be fairly flexible if you have a different workflow. The main thing is getting the icon unicode points out of flat css and into a map, which can then be iterated thru in various interesting ways.